React Fullstack API with Authentication

+++ app/Http/Controllers/PostController.php
@@ -3,22 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
-use App\Http\Requests\StorePostRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
         return Post::all();
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         $validate = $request->validate([
@@ -26,24 +20,21 @@ class PostController extends Controller
             'body' => 'required'
         ]);
 
-        $post = Post::create($validate);
+        $post = $request->user()->posts()->create($validate);
 
         return $post;
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Post $post)
     {
         return $post;
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Post $post)
     {
+        // 🔥 Gate
+        Gate::authorize('modify', $post);
+
         $validate = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required'
@@ -54,11 +45,11 @@ class PostController extends Controller
         return $post;
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Post $post)
     {
+        // 🔥 Gate
+        Gate::authorize('modify', $post);
+
         $post->delete();
 
         return ["message" => "Post Was Deleted!"];
